refactor: Rename to initIpcListeners
